Rename /util.js to /util
