+++ src/util/filterNotes.js
